+++ db/migrate/20260416182139_add_params_to_rails_url_shortener_visits.rb
@@ -1,0 +1,5 @@
+class AddParamsToRailsUrlShortenerVisits < ActiveRecord::Migration[7.2]
+  def change
+    add_column :rails_url_shortener_visits, :params, :text
+  end
+end